services/order-service/requirements.txt
Flask==2.2.5
requests==2.31.0
psycopg2-binary==2.9.7
